@@ -1,11 +1,11 @@
-// swift-tools-version:5.2
+// swift-tools-version:5.3
 
 import PackageDescription
 
 let package = Package(
   name: "CryptoSwift",
   platforms: [
-    .macOS(.v10_12), .iOS(.v9), .tvOS(.v9)
+    .macOS(.v10_12), .iOS(.v9), .tvOS(.v9), .watchOS(.v2)
   ],
   products: [
     .library(
