@@ -23,13 +23,16 @@ public struct CBC: BlockMode {
   }
 
   public let options: BlockModeOption = [.initializationVectorRequired, .paddingRequired]
+  
   private let iv: Array<UInt8>
+
+  public let customBlockSize: Int? = nil
 
   public init(iv: Array<UInt8>) {
     self.iv = iv
   }
 
-  public func worker(blockSize: Int, cipherOperation: @escaping CipherOperationOnBlock) throws -> CipherModeWorker {
+  public func worker(blockSize: Int, cipherOperation: @escaping CipherOperationOnBlock, encryptionOperation: @escaping CipherOperationOnBlock) throws -> CipherModeWorker {
     if self.iv.count != blockSize {
       throw Error.invalidInitializationVector
     }
@@ -45,12 +48,14 @@ struct CBCModeWorker: BlockModeWorker {
   private let iv: ArraySlice<UInt8>
   private var prev: ArraySlice<UInt8>?
 
+  @inlinable
   init(blockSize: Int, iv: ArraySlice<UInt8>, cipherOperation: @escaping CipherOperationOnBlock) {
     self.blockSize = blockSize
     self.iv = iv
     self.cipherOperation = cipherOperation
   }
 
+  @inlinable
   mutating func encrypt(block plaintext: ArraySlice<UInt8>) -> Array<UInt8> {
     guard let ciphertext = cipherOperation(xor(prev ?? iv, plaintext)) else {
       return Array(plaintext)
@@ -59,6 +64,7 @@ struct CBCModeWorker: BlockModeWorker {
     return ciphertext
   }
 
+  @inlinable
   mutating func decrypt(block ciphertext: ArraySlice<UInt8>) -> Array<UInt8> {
     guard let plaintext = cipherOperation(ciphertext) else {
       return Array(ciphertext)
