@@ -14,6 +14,7 @@
 //
 extension Collection where Self.Element == UInt8, Self.Index == Int {
   // Big endian order
+  @inlinable
   func toUInt32Array() -> Array<UInt32> {
     guard !isEmpty else {
       return []
@@ -33,6 +34,7 @@ extension Collection where Self.Element == UInt8, Self.Index == Int {
   }
 
   // Big endian order
+  @inlinable
   func toUInt64Array() -> Array<UInt64> {
     guard !isEmpty else {
       return []
@@ -52,7 +54,8 @@ extension Collection where Self.Element == UInt8, Self.Index == Int {
   }
 }
 
-private func strideCount(from: Int, to: Int, by: Int) -> Int {
+@usableFromInline
+func strideCount(from: Int, to: Int, by: Int) -> Int {
     let count = to - from
     return count / by + (count % by > 0 ? 1 : 0)
 }
