@@ -15,13 +15,15 @@
 //
 
 #if canImport(Darwin)
-  import Darwin
-#else
-  import Glibc
+import Darwin
+#elseif canImport(Glibc)
+import Glibc
+#elseif canImport(ucrt)
+import ucrt
 #endif
 
 extension FixedWidthInteger {
-  @_transparent
+  @inlinable
   func bytes(totalBytes: Int = MemoryLayout<Self>.size) -> Array<UInt8> {
     arrayOfBytes(value: self.littleEndian, length: totalBytes)
     // TODO: adjust bytes order
