@@ -18,9 +18,11 @@
 //
 
 #if canImport(Darwin)
-  import Darwin
-#else
-  import Glibc
+import Darwin
+#elseif canImport(Glibc)
+import Glibc
+#elseif canImport(ucrt)
+import ucrt
 #endif
 
 public final class SHA3: DigestType {
@@ -37,8 +39,12 @@ public final class SHA3: DigestType {
   public let digestLength: Int
   public let markByte: UInt8
 
-  fileprivate var accumulated = Array<UInt8>()
-  fileprivate var accumulatedHash: Array<UInt64>
+  @usableFromInline
+  var accumulated = Array<UInt8>()
+
+
+  @usableFromInline
+  var accumulatedHash: Array<UInt64>
 
   public enum Variant {
     case sha224, sha256, sha384, sha512, keccak224, keccak256, keccak384, keccak512
@@ -81,6 +87,7 @@ public final class SHA3: DigestType {
     self.accumulatedHash = Array<UInt64>(repeating: 0, count: self.digestLength)
   }
 
+  @inlinable
   public func calculate(for bytes: Array<UInt8>) -> Array<UInt8> {
     do {
       return try update(withBytes: bytes.slice, isLast: true)
@@ -175,7 +182,8 @@ public final class SHA3: DigestType {
     a[0] ^= self.round_constants[round]
   }
 
-  fileprivate func process(block chunk: ArraySlice<UInt64>, currentHash hh: inout Array<UInt64>) {
+  @usableFromInline
+  func process(block chunk: ArraySlice<UInt64>, currentHash hh: inout Array<UInt64>) {
     // expand
     hh[0] ^= chunk[0].littleEndian
     hh[1] ^= chunk[1].littleEndian
@@ -249,6 +257,8 @@ public final class SHA3: DigestType {
 }
 
 extension SHA3: Updatable {
+
+  @inlinable
   public func update(withBytes bytes: ArraySlice<UInt8>, isLast: Bool = false) throws -> Array<UInt8> {
     self.accumulated += bytes
 
