@@ -26,7 +26,7 @@
 
     [progress resignCurrent];
 
-    [self waitForExpectationsWithTimeout:1 handler:nil];
+    [self waitForExpectationsWithTimeout:5 handler:nil];
 }
 
 - (void)testProgressDoesNotExceed100Percent {
