@@ -1,10 +1,10 @@
 import XCTest
 
-import A_
+import APlus
 import CorePromise
 
 var tests = [XCTestCaseEntry]()
-tests += A_.__allTests()
+tests += APlus.__allTests()
 tests += CorePromise.__allTests()
 
 XCTMain(tests)
