@@ -18,25 +18,31 @@ class ViewController: UIViewController {
         super.viewDidLoad()
         getCaseCount()
         //getTestInfo(1)
+        //runTest(304)
     }
     
     func removeSocket(_ s: WebSocket?) {
-        socketArray = socketArray.filter{$0 != s}
+        guard let s = s else {return}
+        socketArray = socketArray.filter{$0 !== s}
     }
     
     func getCaseCount() {
-        
-        let s = WebSocket(url: URL(string: "ws://\(host)/getCaseCount")!, protocols: [])
+        let req = URLRequest(url: URL(string: "ws://\(host)/getCaseCount")!)
+        let s = WebSocket(request: req)
         socketArray.append(s)
-        s.onText = { [weak self]  (text: String) in
-            if let c = Int(text) {
-                print("number of cases is: \(c)")
-                self?.caseCount = c
+        s.onEvent = { [weak self] event in
+            switch event {
+            case .text(let string):
+                if let c = Int(string) {
+                    print("number of cases is: \(c)")
+                    self?.caseCount = c
+                }
+            case .disconnected(_, _):
+                self?.runTest(1)
+                self?.removeSocket(s)
+            default:
+                break
             }
-        }
-        s.onDisconnect = { [weak self, weak s]  (error: Error?) in
-            self?.getTestInfo(1)
-            self?.removeSocket(s)
         }
         s.connect()
     }
@@ -44,7 +50,7 @@ class ViewController: UIViewController {
     func getTestInfo(_ caseNum: Int) {
         let s = createSocket("getCaseInfo",caseNum)
         socketArray.append(s)
-        s.onText = { (text: String) in
+//        s.onText = { (text: String) in
 //            let data = text.dataUsingEncoding(NSUTF8StringEncoding)
 //            do {
 //                let resp: AnyObject? = try NSJSONSerialization.JSONObjectWithData(data!,
@@ -60,14 +66,19 @@ class ViewController: UIViewController {
 //                print("error parsing the json")
 //            }
 
-        }
+//        }
         var once = false
-        s.onDisconnect = { [weak self, weak s]  (error: Error?) in
-            if !once {
-                once = true
-                self?.runTest(caseNum)
+        s.onEvent = { [weak self] event in
+            switch event {
+            case .disconnected(_, _), .error(_):
+                if !once {
+                    once = true
+                    self?.runTest(caseNum)
+                }
+                self?.removeSocket(s)
+            default:
+                break
             }
-            self?.removeSocket(s)
         }
         s.connect()
     }
@@ -75,92 +86,99 @@ class ViewController: UIViewController {
     func runTest(_ caseNum: Int) {
         let s = createSocket("runCase",caseNum)
         self.socketArray.append(s)
-        s.onText = { [weak s]  (text: String) in
-            s?.write(string: text)
-        }
-        s.onData = { [weak s]  (data: Data) in
-            s?.write(data: data)
-        }
+        
         var once = false
-        s.onDisconnect = {[weak self, weak s] (error: Error?) in
-            if !once {
-                once = true
-                print("case:\(caseNum) finished")
-                //self?.verifyTest(caseNum) //disabled since it slows down the tests
-                let nextCase = caseNum+1
-                if nextCase <= (self?.caseCount)! {
-                    self?.runTest(nextCase)
-                    //self?.getTestInfo(nextCase) //disabled since it slows down the tests
-                } else {
-                    self?.finishReports()
+        s.onEvent = { [weak self, weak s] event in
+            switch event {
+            case .disconnected(_, _), .error(_):
+                if !once {
+                    once = true
+                    print("case:\(caseNum) finished")
+                    //self?.verifyTest(caseNum) //disabled since it slows down the tests
+                    let nextCase = caseNum+1
+                    if nextCase <= (self?.caseCount)! {
+                        self?.runTest(nextCase)
+                        //self?.getTestInfo(nextCase) //disabled since it slows down the tests
+                    } else {
+                        self?.finishReports()
+                    }
+                    self?.removeSocket(s)
                 }
                 self?.removeSocket(s)
+            case .text(let string):
+               s?.write(string: string)
+            case .binary(let data):
+               s?.write(data: data)
+//            case .error(let error):
+//                print("got an error: \(error)")
+            default:
+                break
             }
         }
         s.connect()
     }
     
-    func verifyTest(_ caseNum: Int) {
-        let s = createSocket("getCaseStatus",caseNum)
-        self.socketArray.append(s)
-        s.onText = { (text: String) in
-            let data = text.data(using: String.Encoding.utf8)
-            do {
-                let resp: Any? = try JSONSerialization.jsonObject(with: data!,
-                    options: JSONSerialization.ReadingOptions())
-                if let dict = resp as? Dictionary<String,String> {
-                    if let status = dict["behavior"] {
-                        if status == "OK" {
-                            print("SUCCESS: \(caseNum)")
-                            return
-                        }
-                    }
-                    print("FAILURE: \(caseNum)")
-                }
-            } catch {
-               print("error parsing the json")
-            }
-        }
-        var once = false
-        s.onDisconnect = { [weak self, weak s]  (error: Error?) in
-            if !once {
-                once = true
-                let nextCase = caseNum+1
-                print("next test is: \(nextCase)")
-                if nextCase <= (self?.caseCount)! {
-                    self?.getTestInfo(nextCase)
-                } else {
-                    self?.finishReports()
-                }
-            }
-            self?.removeSocket(s)
-        }
-        s.connect()
-    }
+//    func verifyTest(_ caseNum: Int) {
+//        let s = createSocket("getCaseStatus",caseNum)
+//        self.socketArray.append(s)
+//        s.onText = { (text: String) in
+//            let data = text.data(using: String.Encoding.utf8)
+//            do {
+//                let resp: Any? = try JSONSerialization.jsonObject(with: data!,
+//                    options: JSONSerialization.ReadingOptions())
+//                if let dict = resp as? Dictionary<String,String> {
+//                    if let status = dict["behavior"] {
+//                        if status == "OK" {
+//                            print("SUCCESS: \(caseNum)")
+//                            return
+//                        }
+//                    }
+//                    print("FAILURE: \(caseNum)")
+//                }
+//            } catch {
+//               print("error parsing the json")
+//            }
+//        }
+//        var once = false
+//        s.onDisconnect = { [weak self, weak s]  (error: Error?) in
+//            if !once {
+//                once = true
+//                let nextCase = caseNum+1
+//                print("next test is: \(nextCase)")
+//                if nextCase <= (self?.caseCount)! {
+//                    self?.getTestInfo(nextCase)
+//                } else {
+//                    self?.finishReports()
+//                }
+//            }
+//            self?.removeSocket(s)
+//        }
+//        s.connect()
+//    }
     
     func finishReports() {
         let s = createSocket("updateReports",0)
         self.socketArray.append(s)
-        s.onDisconnect = { [weak self, weak s]  (error: Error?) in
-            print("finished all the tests!")
-            self?.removeSocket(s)
+        s.onEvent = { [weak self, weak s] event in
+            switch event {
+            case .disconnected(_, _):
+                print("finished all the tests!")
+                self?.removeSocket(s)
+            default:
+                break
+            }
         }
         s.connect()
     }
     
     func createSocket(_ cmd: String, _ caseNum: Int) -> WebSocket {
-        return WebSocket(url: URL(string: "ws://\(host)\(buildPath(cmd,caseNum))")!, protocols: [])
+        let req = URLRequest(url: URL(string: "ws://\(host)\(buildPath(cmd,caseNum))")!)
+        //return WebSocket(request: req, compressionHandler: WSCompression())
+        return WebSocket(request: req)
     }
     
     func buildPath(_ cmd: String, _ caseNum: Int) -> String {
         return "/\(cmd)?case=\(caseNum)&agent=Starscream"
     }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-
 }
 
